Doodle combining tilings and poisson-disc.
Started experimenting with an idea to create a piece which combines poisson-disc sampling and plane tilings. At the moment I have moved the poisson-disk sampling code I developed previously to a library file, I have cleaned it and I use it to sample a rectangle in the centre of the screen.